+++ beaker/data_model/__init__.py
@@ -1,5 +1,4 @@
 from .account import *
-from .base import *
 from .cluster import *
 from .dataset import *
 from .experiment import *